+++ tests/test_compat_context.py
@@ -1,5 +1,6 @@
 """Test that the "compatible" render context is more compatible with Ruby Liquid
 than the default render context."""
+
 import asyncio
 from unittest import TestCase
 
@@ -15,7 +16,8 @@ class CompatContextTestCase(TestCase):
     def test_string_indices(self) -> None:
         """Test that we can not access characters in a string by index."""
         env = Environment()
-        context = FutureContext(env, globals={"some": "thing"})
+        template = env.from_string("Hello, {{ some }}")
+        context = FutureContext(template, globals={"some": "thing"})
         self.assertTrue(is_undefined(context.get(["some", 1])))
 
         with self.subTest(msg="async"):
@@ -24,7 +26,8 @@ class CompatContextTestCase(TestCase):
     def test_special_properties(self) -> None:
         """Test special `size`, `first` and last properties."""
         env = Environment()
-        context = FutureContext(env, globals={"some": "thing"})
+        template = env.from_string("Hello, {{ some }}")
+        context = FutureContext(template, globals={"some": "thing"})
         self.assertEqual(context.get(["some", "size"]), 5)
         self.assertTrue(is_undefined(context.get(["some", "first"])))
         self.assertTrue(is_undefined(context.get(["some", "last"])))
@@ -41,7 +44,8 @@ class CompatContextTestCase(TestCase):
     def test_cycle_without_name(self) -> None:
         """Test unnamed cycles."""
         env = Environment()
-        context = FutureContext(env)
+        template = env.from_string("Hello, {{ some }}")
+        context = FutureContext(template)
         self.assertEqual(context.cycle("", [1, 2, 3]), 1)
         self.assertEqual(context.cycle("", ["x", "y", "z"]), "x")
         self.assertEqual(context.cycle("", [1, 2, 3]), 2)
@@ -49,7 +53,8 @@ class CompatContextTestCase(TestCase):
     def test_cycle_wraps(self) -> None:
         """Test that we wrap around."""
         env = Environment()
-        context = FutureContext(env)
+        template = env.from_string("Hello, {{ some }}")
+        context = FutureContext(template)
         self.assertEqual(context.cycle("", ["some", "other"]), "some")
         self.assertEqual(context.cycle("", ["some", "other"]), "other")
         self.assertEqual(context.cycle("", ["some", "other"]), "some")
@@ -57,7 +62,8 @@ class CompatContextTestCase(TestCase):
     def test_named_cycle_groups(self) -> None:
         """Test that named cycles ignore arguments."""
         env = Environment()
-        context = FutureContext(env)
+        template = env.from_string("Hello, {{ some }}")
+        context = FutureContext(template)
         self.assertEqual(context.cycle("a", [1, 2, 3]), 1)
         self.assertEqual(context.cycle("a", ["x", "y", "z"]), "y")
         self.assertEqual(context.cycle("a", [1, 2, 3]), 3)
@@ -65,7 +71,8 @@ class CompatContextTestCase(TestCase):
     def test_named_cycles_with_shrinking_lengths(self) -> None:
         """Test that we handle cycles with a shrinking number of arguments."""
         env = Environment()
-        context = FutureContext(env)
+        template = env.from_string("Hello, {{ some }}")
+        context = FutureContext(template)
         self.assertEqual(context.cycle("a", [1, 2, 3]), 1)
         self.assertEqual(context.cycle("a", ["x", "y"]), "y")
         self.assertEqual(context.cycle("a", ["x", "y"]), "x")
@@ -73,7 +80,8 @@ class CompatContextTestCase(TestCase):
     def test_named_cycles_with_growing_lengths(self) -> None:
         """Test that we handle cycles with a growing number of arguments."""
         env = Environment()
-        context = FutureContext(env)
+        template = env.from_string("Hello, {{ some }}")
+        context = FutureContext(template)
         self.assertEqual(context.cycle("a", [1, 2]), 1)
         self.assertEqual(context.cycle("a", ["x", "y", "z"]), "y")
         self.assertEqual(context.cycle("a", ["x", "y", "z"]), "z")