+++ tests/test_context_loader.py
@@ -1,8 +1,8 @@
 import asyncio
 
-from liquid import Context
 from liquid import DictLoader
 from liquid import Environment
+from liquid import RenderContext
 from liquid.loaders import TemplateSource
 
 
@@ -12,14 +12,14 @@ class MockContextLoader(DictLoader):
         super().__init__(templates)
 
     def get_source_with_context(
-        self, context: Context, template_name: str, **kwargs: str
+        self, context: RenderContext, template_name: str, **kwargs: str
     ) -> TemplateSource:
         self.kwargs.update(kwargs)
         self.kwargs["uid"] = context.resolve("uid", default=None)
         return super().get_source(context.env, template_name)
 
     async def get_source_with_context_async(
-        self, context: Context, template_name: str, **kwargs: str
+        self, context: RenderContext, template_name: str, **kwargs: str
     ) -> TemplateSource:
         self.kwargs.update(kwargs)
         self.kwargs["uid"] = context.resolve("uid", default=None)